+++ src/pages/AllCampsite/index.js
@@ -2,8 +2,8 @@ import React, {useEffect, useState} from "react";
 import axios from "axios";
 // Components
 import Header from "../../components/Header";
-import {SpotBasedSearch} from "../../components/GoCampingAPI/index";
-import AreacodeAPI from "../../components/AreacodeAPI";
+import {SpotBasedSearch} from "../../components/API/GoCampingAPI/index";
+import AreacodeAPI from "../../components/API/AreacodeAPI";
 import OneCampsiteOnList from "../../components/OneCampsiteOnList";
 import Pagination from "../../components/Pagination";
 import Footer from "../../components/Footer";
